+++ chess.py
@@ -1,5 +1,5 @@
 from board import Board
-from exceptions import InvalidMove
+from exceptions import InvalidMove,EmptyPosition,InvalidTurn
 
 class Chess:
     def __init__(self):
@@ -10,10 +10,14 @@ class Chess:
         return True
 
     def move(self,from_row,from_col,to_row,to_col):
-        self.change_turn()
-        piece=self.__board__.get_piece(from_row,from_col)
-        if piece.valid_positions(from_row,from_col,to_row,to_col):
+        piece = self.__board__.get_piece(from_row, from_col)
+        if not piece:
+            raise EmptyPosition()
+        if not piece.get_color() == self.__turn__:
+            raise InvalidTurn()
+        if not piece.valid_positions(from_row, from_col, to_row, to_col):
             raise InvalidMove()
+        self.__board__.move(from_row, from_col, to_row, to_col)
         self.change_turn()
     
     def turn(self):