+++ exceptions.py
@@ -1,8 +1,16 @@
 class InvalidMove(Exception):
-    pass
+    message = "Movimieto de pieza invalido"
+    def __str__(self):
+        return self.message
 
-class InvalidMoveNoPiece(InvalidMove):
-    pass
+class InvalidTurn(InvalidMove):
+    message = "No puedes mover pieza de otro jugador"
+
+class EmptyPosition(InvalidMove):
+    message = "La posicion esta vacia"
 
 class InvalidMoveRookMove(InvalidMove):
-    pass
+    message='La torre no se puede mover en esa direccion'
+
+class OutOfBoard(InvalidMove):
+    message = "La posicion indicada se encuentra fuera del tablero"