+++ pieces.py
@@ -8,6 +8,9 @@ class Piece:
             return self.white_str
         else: return self.black_str
         
+    def get_color(self):
+        return self.__color__
+    
 class Pawn(Piece):
     def __init__(self,color,board):
         super().__init__(color,board)