+++ tests/test_board.py
@@ -2,6 +2,7 @@ import unittest
 from board import Board
 from pieces import Pawn
 from rook import Rook
+from exceptions import OutOfBoard
 
 class TestBoard(unittest.TestCase):
     def setUp(self):
@@ -104,6 +105,39 @@ class TestBoard(unittest.TestCase):
         self.assertIsNone(self.board.get_piece(5,5))
         self.assertIsNone(self.board.get_piece(5,6))
         self.assertIsNone(self.board.get_piece(5,7))
+
+    def test_move(self):
+        board=Board(for_test=True)
+        rook=Rook(color='BLACK', board=board)
+        board.set_piece(0, 0, rook)
+        board.move(0,0,0,1)
+
+        self.assertIsInstance(
+            board.get_piece(0, 1),
+            Rook)
+        
+        self.assertEqual(
+            str(board),
+            (
+                " ♖      \n"
+                "        \n"
+                "        \n"
+                "        \n"
+                "        \n"
+                "        \n"
+                "        \n"
+                "        \n"
+            )
+        )
+
+    def test_get_piece_out_of_range(self):
+        board = Board(for_test=True)
+        with self.assertRaises(OutOfBoard) as exc:
+            board.get_piece(10, 10)
+        self.assertEqual(
+            exc.exception.message,
+            "La posicion indicada se encuentra fuera del tablero"
+        )
 
 if __name__=='__main__':
     unittest.main()