+++ tests/test_rook.py
@@ -46,6 +46,12 @@ class TestRook(unittest.TestCase):
             possibles,
             [(5, 1), (6, 1)]
         )
-
+    
+    def test_move_diagonal_desc(self):
+        board=Board()
+        rook=board.get_piece(0,0)
+        is_possible=rook.valid_positions(0,0,1,1)
+        self.assertFalse(is_possible)
+        
 if __name__=='__main__':
     unittest.main()